feat(apm/version): bump version based on latest published version on chain
This commit implements #196 to lay the foundation for #195. As discussed in the
issues, the goal is to have

```
$ aragon apm version
```

to not just update the version number in an aragon app's manifest file based on
the currently stored version, but instead reach out to the latest published
version on the chain to increase from there.

This functionality is now exposed as a new API as well so it can be used by
other commands, which is required to implement #195.

Closes #195