feat(apm/publish): introduce version-bump option
This allows users to increase their app's version number straight from the

```
$ aragon apm publish
```

command as discussed in #195. Possible value for this option are:

- patch
- minor
- major

This commit also includes an optimization where an app's contracts are not
compiled and deployed in case the version to be published is a patch or minor
update. As only major updates are allowed for newly deployed contracts.

Closes #195